requirements.txt
python-dotenv>=1.0.0
feedparser>=6.0.10
requests>=2.31.0
python-dateutil>=2.8.2
anthropic>=0.39.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
schedule>=1.2.0
moviepy>=2.1.0
imageio-ffmpeg>=0.5.0
edge-tts>=6.1.0
Pillow>=10.0.0
numpy>=1.24.0
tweepy>=4.14.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
